Product Owner qualities suggestion
Since a product owner is responsible for shaping the work of the development team and ensuring they only implement the necessary features in the time allocated for the sprint, they are best suited for this task by having development experience and an understanding of the technologies in use. If they don't then they may struggle to communicate with the development team or they may not foresee traps laying on the road ahead. There will  be exceptions to this of course, but in general these qualities would only increase the odds of success.